+++ tablet_qt/tasklib/inittasks.cpp
@@ -99,6 +99,7 @@
 
 #include "tasks/panss.h"
 #include "tasks/patientsatisfaction.h"
+#include "tasks/pcl5.h"
 #include "tasks/pclc.h"
 #include "tasks/pclm.h"
 #include "tasks/pcls.h"
@@ -215,6 +216,7 @@ void InitTasks(TaskFactory& factory)
 
     initializePanss(factory);
     initializePatientSatisfaction(factory);
+    initializePcl5(factory);
     initializePclC(factory);
     initializePclM(factory);
     initializePclS(factory);